Missing pidgin pkg-config files set to warn instead of error during configure
Users running custom builds might not be using pkg-config files but could still
specify the necessary environment variables for compiling to succeed.  The
configure script will now print a warning with a note about these variables if
it can't find pidgin.pc, and it will no longer error and stop the build.