Reset the Buddy List window icon in addition to its title when necessary
When all conversations are closed (or the plugin is disabled), show the initial
Buddy List window icon instead of the previous conversation's.  Since the Buddy
List does not specify its own icon list, it uses the default list.  This allows
the window's icon list to simply be unset for it to display the defaults again.

I've been using an icon-free window manager for so long that I forgot Pidgin
changes them...