Skip some bindings so key events make it to conversation signal handlers
The signal handler gtk_blist_window_key_press_cb in gtkblist.c will intercept
and stop key events on the Buddy List window if there is a matching key binding
on the focused widget.  This poses a problem for a few key events that are
manually processed in gtkconv.c's signal handlers.

This change ignores bindings on a few key events important to conversations.
For example, Control+Up retrieves from message history in a conversation, but
has default binding to move the cursor up one paragraph.  (How many people even
know that exists?)  More important default bindings, such as Control+Tab to
change focus, are not modified.

Hopefully this will be a temporary change in the long run, and something can be
done with gtk_blist_window_key_press_cb in Pidgin to support these cases.