Support directly building a Windows DLL without requiring Pidgin's source
These tweaks to the autoconf/automake files will generate a DLL when configure
is run for a MinGW build environment.  If the MinGW libraries libpurple.dll and
libpidgin.dll are properly installed (along with their headers and preferably
pkg-config files), then the plugin can be built without the Pidgin source tree.