Enable menu accelerators from the conversation window on the Buddy List
There are a couple accelerators from the conversation window that overlap those
on the Buddy List (e.g. Control+L), but the behavior of these items in my tests
was intuitive enough as to not warrant reassigning accelerators (which I really
don't want to do, changing standard keys etc.).  It seems the conversations'
menu action is taken when a conversation is open, and the Buddy List's action
is taken otherwise.  Well, that's nice.