Attempt to standardize on user-visible strings for translation purposes
I was being sloppy (sloppier) about marking strings for translation, so these
changes should rectify that.  The important change is making global definitions
for strings that appear in multiple places (and just directly embedding any of
the string definitions that appear only once, to be consistent).  All strings
that are visible to the user should now be properly marked with the typical "_"
and "N_" keyword macros.