Add a fallback option to the conversation placement function
If the conversation placement function is called while the Buddy List is not
fused with a conversation window, lots of pain would result.  This will now use
the "last opened window" placement function (the Pidgin default) if the Buddy
List is not merged.

I didn't realize that the placement function is still used even when the plugin
is disabled (despite deregistering everything).  That oversight probably caused
lots of problems for anyone trying to disable the plugin without restarting.