Preserve the original conversation window for when the windows are split
The conversation GtkWindow should now remain hidden during the entire time the
plugin is enabled, rather than freeing it immediately.  This window is simply
reconstructed from its old parts and displayed when the plugin is disabled.  It
should automatically destroy itself if the plugin is disabled while there are
no conversations open in it.