Supply a script to build a Windows installer
I'm already providing an RPM spec, so why not?  This can be executed directly
as a script on a Unix-like system with makensis installed, or makensis can run
on it in a more traditional manner.  (The installer will only build once "make"
has succeeded at building the plugin.)

It simply looks at the Pidgin registry entry for its installation path (if it's
defined and still valid), or otherwise defaults to %ProgramFiles%\Pidgin.  With
that path, after verifying pidgin.exe is indeed in there, it writes the plugin
DLL to the plugins folder and creates an uninstaller at the main path.

For now, this is probably only useful for users who are unable to locate their
%AppData%\.purple\plugins folder.  However, in the future, it could simplify
installing and removing translations etc. in addition to the plugin itself.