Display the software license in the Windows installer
This commit is more of a complete rewrite than the subject suggests, but I do
not want a polluted git history because of this file.  Most changes are due to
my paying closer attention to NSIS documentation, such as it explicitly lists
pages now because implicit pages were deprecated.  The script's layout was also
reorganized because this is a censored version of the file from my local branch
with translations and a second plugin.  Those sections will now add cleanly in
the future if I ever get around to distributing that stuff.