Exclude "Send To" from the migrated menu items
This menu item is destroyed and recreated every time the active conversation
tab is changed.  It is then hard-coded into the third menu position, which will
mess up the ordering of Buddy List menu items.  The least confusing solution is
to simply not include the menu in the merged window until its behavior is fixed
in Pidgin, or to wait until Pidgin is ported to GTK >= 3.2 when a signal is
finally implemented for when items are added to menus.