Add support for building translations
This only includes the pieces necessary to build the PO files and does not
include any PO files themselves.  When the project is built, it will create the
POT file under the "po" directory.  This can be run through msginit to write a
PO file.  The language code can then be appended to "po/LINGUAS", and then the
translation file will be built after a reconfigure.