Explicitly detect the system's Pidgin plugin directory
This will have different behavior depending on the pkg-config version
of the system where distribution files are generated.  (I wrote a macro
included in pkg-config 0.28, which is still fairly new and uncommon, so
this change needed a fallback test.)

If the system uses pkg-config older than 0.28, it should preserve the
existing behavior.  The configure script will define plugindir to be
'${libdir}/pidgin' (Pidgin's default), and it can be overridden on the
make command line (i.e. "make install plugindir=~/.purple/plugins").

If the system uses pkg-config with version at least 0.28, a configure
variable pidgin_plugindir is declared precious.  It will be set to the
plugindir value from Pidgin's pkg-config settings if it was not already
overridden by environment variables.  (This is the same as pidgin_LIBS
and pidgin_CFLAGS.)  It will default to '${libdir}/pidgin' again if the
pkg-config calls fail.  The resulting value of pidgin_plugindir is used
as plugindir in the Makefile, which can be overriden as above.